Create Critical Connection
Problem statement
You are given a network with ‘N’ system nodes [0 to N - 1] and ‘M’ connection. Your task is to find out all critical connections in a given network.

Note: A connection between node ‘u’ and ‘v’ is said to be a critical connection, if after removal of a connection ‘u’ - ‘v’, there is no connection between node ‘u’ and ‘v’ and the network goes down.

For example:

For given N = 4, M = 4, 

The connection between system node 0 and 1 is a critical connection.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 5
1 <= N, M <= 10 ^ 6
0 <= u, v <= N - 1
There are no repeated connections. 

Time Limit: 1 sec.
Sample Input 1:
2
4 4
0 1
1 2
1 3
2 3
6 6
0 1
1 2
1 3
3 4
4 5
0 4
Sample Output 1:
1
0 1
2
1 2
4 5
Explanation of Sample Output1:
In the first test case, the given network  looks like this: 

The only connection between nodes 0 and 1 is critical. So, the number of critical connections is 1. 

In the second test case, the given network looks like this: 

The connection between nodes 1, 2, and 4, 5 is a critical connection. So, there are 2 critical connections. If anyone's failure makes the network go down
Sample Input 2:
2
5 6
0 1
1 3
2 1
2 2
2 4
3 3
5 3
0 1
2 0
4 3
Sample Output 2:
4
0 1
1 2
1 3
2 4
3
0 1
0 2
3 4