Create Mailing Problem
Problem statement
Dr. Doofenshmirtz is trying to make another big mischief. To prevent this, Major Francis Monogram wants to send mail to Perry the Platypus to inform him about the mischief. He just needs to send a secret message ‘s’ to perry. He has a special keyboard “keyboard” in which all the letters are in a straight line but jumbled. The time taken to move his finger from index ‘i’ to index ‘j’ is |j - i|. He wants to know how much time will it take to send the secret message to Perry.

Note :
Initially, he is at the first letter of the keyboard.

All letters are in lower-case English letters. 
For Example :
Let s = “perry”, keyboard = “qwertyuiopasdfghjklzxcvbnm”.

Now In this example, Francis will start from the first index and go to ‘p’, which will take 9 seconds, now from ‘p’ to ‘e’, the distance is of 7 indexes, from ‘e’ to ‘r’ the distance is 1, from ‘r’ to ‘r’ it will be zero, and from ‘r’ to ‘y’ the distance will be 2. Hence the total time taken is 9 + 7 + 1 + 0 + 2 = 19.
Hence the answer is 19.
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 100
keyboard.length = 26
1 <= |S| <= 10^5, where |S| represents the length of the String S.

Time limit: 1 sec
Sample Input 1 :
2
abcdefghijklmnopqrstuvwxyz
abc
zyxwvutsrqponmlkjihgfedcba
zayb
Sample Output 1 :
2
72
Explanation For Sample Output 1 :
For the first test case, Francis is currently on ‘a’ so he does not have to move his fingers for the first letter and then move one index from ‘a’ to ‘b’ and again one index from ‘b’ to ‘c’.
Hence the answer is 2.

For the second test case, he will start from the 0th index and then move to 25 to type ‘a’, then to index 1 to type ‘y’, and at last to index 24 to type ‘b’. Hence the total time taken is 0 + 25 + 24 + 23 = 72.
Sample Input 2 :
2
mvpglaejurowcfbtisdkzhxqny
ferb
bjqtcxmkgzfeowdlsuaypnivrh
phineas
Sample Output 2 :
28
48