Create Minimum Difference Of Subarrays
Problem statement
You are given an array 'ARR' that consists of ‘N’ numbers. Your task is to split the array into two non-empty subarrays such that the absolute difference between their sum is minimum. For example,
If the given array is [6,7,1,8,5]. We can split the array into [6,7,1] and [8,5].
Hence the answer will be |14-13| = 1 which is minimum.
Detailed explanation ( Input/output format, Notes, Images )
Constraints:
1 <= T <= 10
1 <= N <= 10^6
-10^3 <= 'ARR'[i] <= 10^3

Time limit: 1 sec
Sample Input 1:
2
5 
6 7 1 8 5
3
6 6 6    
Sample Output 1:
1
6
Explanation of sample input 1:
For the first test case, 
We can split the array into [6,7] and [1,8,5]. The absolute difference is  |13-14| =1.
Hence,  the answer is 1.

For the second test case,
We can split the array into [6] and [6,6]. The absolute difference is |6-12| =6.
Hence, the answer is 6.    
Sample Input 2:
2
5
4 1 20 9 9
4
7 9 8 10
Sample Output 2:
7
2