Create NINJA'S JUMP
Problem statement
Ninja is assigned a task to reach the last stone by his master. These stones are numbered with some value and in the form of an array. He is allowed to jump either odd-numbered jumps or even-numbered jumps and has to reach the last stone.

So your task is to find the number of starting index from which he may start jumping so he reaches the last stones. You are provided with the given array and you have to find the number of starting index of the array from which Ninja can reach the end of the array by jumping some number of times.

For jumping you have to follow below instructions:

You may jump forward from index ‘i’ to index ‘j’ (with i < j) in the following way:

During odd-numbered jumps (i.e., jumps 1, 3, 5, ...), you jump to the index ‘j’ such that ‘arr[i] <= arr[j]’ and ‘arr[j]’ is the smallest possible value. If there are multiple such indices ‘j’, you can only jump to the smallest such index j.

During even-numbered jumps (i.e., jumps 2, 4, 6, ...), you jump to the index ‘j’ such that ‘arr[i] >= arr[j]’ and ‘arr[j]’ is the largest possible value. If there are multiple such indices ‘j’, you can only jump to the smallest such index ‘j’.

It may be the case that for some index ‘i’, there are no legal jumps.

Detailed explanation ( Input/output format, Notes, Images )
Sample Input 1:
2
5
20 23 22 24 25
5
5 6 4 4 7
Sample Output 1:
2
3
Explanation for sample input 1:
Test Case 1:
For the first test case, if we start from index i = 0, we can make our first jump to i = 2, since it is smaller than all other array values but then we cannot jump anymore. 
If we start from the index i = 1, or i = 2 we can make the first jump to i = 3 then we cannot jump anymore.
If we start from index i = 3 we have to make only one jump and we reach the end.
And if we start from index i = 4 we are already at the end.
So we return ‘2’ as our answer as there are ‘2’ good starting index.

Test Case 2:
For the test case if we start From starting index i = 0, we make jumps to i = 1, i = 2, i = 3
From starting index i = 1, we jump to i = 4, so we reach the end.
From starting index i = 2, we jump to i = 3, and then we can't jump anymore.
From starting index i = 3, we jump to i = 4, so we reach the end.
From starting index i = 4, we are already at the end.

In total, there are 3 different starting indices i = 1, i = 3, and i = 4, where we can reach the end with some number of jumps so we return ‘3’ as our answer.
Sample Input 2:
1
5
5 1 3 4 2
Sample Output 2:
3
Explanation of sample input 2:
As we can reach the end by starting from index 1, 2, 4