Create Loudspeakers
Problem statement
The famous cultural festival of your locality is near and as a member of the committee, you are given the responsibility of arranging loudspeakers for the cultural festival that has to take place. Now, you have arranged N loud-speakers where the ith loudspeaker has loudness or amplitude LOUD[i]. Now when one loudspeaker of loudness Y is kept with another loudspeaker of loudness X, the combined loudness is expressed as Y^X.

So, when N loudspeakers are kept in order A1, A2, A3……….An, the combined loudness will be (A1^(A2^(A3^(...........(An))))). So, your task is to rearrange the loudspeakers such that the combined loudness emitted is maximum.

Example:-
Let, 
N = 3
LOUD = [4, 5, 6]
Answer:- [4, 5, 6]
Explanation:- The answer should be 4,5,6 because out of all the rearrangements that are possible the array [4,5,6] will emit the largest combined loudness. Arrangement 4,5,6 emits an amplitude of 4^5^6 (1152921504606846976) and arrangement 6,5,4 emits an amplitude of 6^5^4 (3656158440062976).
Detailed explanation ( Input/output format, Notes, Images )
Constraints :
1 <= T <= 5
1 <= N <= 10^5
1 <= LOUD[i] <= 10^9

Note:- Loudspeakers having an amplitude of more than 1 will all have unique amplitudes.

Time Limit = 1 sec
Sample Input 1 :
2
2
2 3
2
2 4
Sample Output 1 :
3 2
2 4
Explanation for Sample Output 1 :
In the first test case, the answer should be 3, 2 because 3, 2 emits a loudness of 9 which is the maximum possible.
In the second test case, the answer should be 2, 4 because 2, 4 emits a loudness of 16 which is the largest possible. (Note that 4, 2 also yields the same output but 2, 4 is lexicographically smaller).
Sample Input 2 :
1
1
7
Sample Output 2 :
7